refactor(direct-control): extract turn completion owner
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only getCiv7TurnCompletionStatus, sendCiv7TurnComplete, sendCiv7TurnUnready, the generated status command, and private turn-completion fallback classifier helpers into src/play/turn-completion.ts while keeping public facade exports in index.ts.

Preserves approval-first send/unready behavior, guard-first status reads, stale notification fallback classification, command strings, parse label, and action result shape. Autoplay source ownership, hotseat runtime proof, AI ingestion, CLI semantic projection, telemetry, Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance remain pending.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- autoplay-and-turn.test.ts
- bun run --cwd packages/cli test -- game.play.end-turn.test.ts
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- git diff --check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- ownership/no-cycle scans for turn-completion symbols and executable facade imports

Local package/source relocation proof only; no runtime, live-game, hotseat product-path, AI ingestion, semantic CLI, telemetry, or procedure-core proof claimed.
