+++ packages/civ7-direct-control/src/index.ts
@@ -58,6 +58,11 @@ import {
 } from "./runtime/tuner-health.js";
 import { getCiv7PlayableStatus as getCiv7PlayableStatusFromModule } from "./runtime/playable-status.js";
 import {
+  getCiv7TurnCompletionStatus as getCiv7TurnCompletionStatusFromModule,
+  sendCiv7TurnComplete as sendCiv7TurnCompleteFromModule,
+  sendCiv7TurnUnready as sendCiv7TurnUnreadyFromModule,
+} from "./play/turn-completion.js";
+import {
   getCiv7NotificationDismissal as getCiv7NotificationDismissalFromModule,
   requestCiv7NotificationDismissal as requestCiv7NotificationDismissalFromModule,
 } from "./play/notifications/dismissal-request.js";
@@ -3435,11 +3440,11 @@ export async function revealCiv7MapForPlayer(
 export async function getCiv7TurnCompletionStatus(
   options: Civ7DirectControlOptions = {},
 ): Promise<Civ7TurnCompletionStatusResult> {
-  const result = await executeCiv7AppUiCommand({
-    ...options,
-    command: buildTurnCompletionStatusCommand(),
+  return await getCiv7TurnCompletionStatusFromModule(options, {
+    executeAppUiCommand: executeCiv7AppUiCommand,
+    parseTurnCompletionStatus: (result, label) =>
+      jsonPayloadFromCommandResult<Civ7TurnCompletionStatusResult>(result, label),
   });
-  return jsonPayloadFromCommandResult<Civ7TurnCompletionStatusResult>(result, "Civ7 turn completion status");
 }
 
 export async function getCiv7PlayNotificationView(
@@ -3482,34 +3487,26 @@ export async function sendCiv7TurnComplete(
   options: Civ7DirectControlOptions = {},
   approval: Civ7ActionApproval,
 ): Promise<Civ7TurnCompletionActionResult> {
-  assertApproved(approval, "sending Civ7 turn complete");
-  const before = await getCiv7TurnCompletionStatus(options);
-  if (probeValue(before.canEndTurn) !== true) {
-    throw new Civ7DirectControlError("command-failed", "Civ7 turn complete is blocked by current game state", {
-      details: before,
-    });
-  }
-  const command = await executeCiv7AppUiCommand({
-    ...options,
-    command: "GameContext.sendTurnComplete()",
+  return await sendCiv7TurnCompleteFromModule(options, approval, {
+    assertApproved,
+    executeAppUiCommand: executeCiv7AppUiCommand,
+    getPlayNotificationView: getCiv7PlayNotificationView,
+    parseTurnCompletionStatus: (result, label) =>
+      jsonPayloadFromCommandResult<Civ7TurnCompletionStatusResult>(result, label),
   });
-  const after = await getCiv7TurnCompletionStatus(options);
-  const verified = probeValue(after.hasSentTurnComplete) === true || probeValue(after.turn) !== probeValue(before.turn);
-  return { before, after, command, verified };
 }
 
 export async function sendCiv7TurnUnready(
   options: Civ7DirectControlOptions = {},
   approval: Civ7ActionApproval,
 ): Promise<Civ7TurnCompletionActionResult> {
-  assertApproved(approval, "sending Civ7 turn unready");
-  const before = await getCiv7TurnCompletionStatus(options);
-  const command = await executeCiv7AppUiCommand({
-    ...options,
-    command: "GameContext.sendUnreadyTurn()",
+  return await sendCiv7TurnUnreadyFromModule(options, approval, {
+    assertApproved,
+    executeAppUiCommand: executeCiv7AppUiCommand,
+    getPlayNotificationView: getCiv7PlayNotificationView,
+    parseTurnCompletionStatus: (result, label) =>
+      jsonPayloadFromCommandResult<Civ7TurnCompletionStatusResult>(result, label),
   });
-  const after = await getCiv7TurnCompletionStatus(options);
-  return { before, after, command, verified: probeValue(after.hasSentTurnComplete) === false };
 }
 
 export async function canStartCiv7UnitOperation(
@@ -4591,29 +4588,6 @@ function buildStopAutoplayCommand(options: Civ7AutoplayOptions): string {
   })()`;
 }
 
-function buildTurnCompletionStatusCommand(): string {
-  return `(() => {
-    ${probeHelperSource()}
-    return JSON.stringify({
-      localPlayerId: GameContext.localPlayerID,
-      turn: probe(() => Game.turn),
-      turnDate: probe(() => Game.getTurnDate()),
-      hasSentTurnComplete: probe(() => GameContext.hasSentTurnComplete()),
-      canEndTurn: probe(() => typeof canEndTurn === "function" ? canEndTurn() : false),
-      blocker: probe(() => typeof Game !== "undefined" && Game.Notifications && typeof Game.Notifications.getEndTurnBlockingType === "function"
-        ? Game.Notifications.getEndTurnBlockingType(GameContext.localPlayerID)
-        : "unknown"),
-      firstReadyUnitId: probe(() => {
-        const id = UI?.Player?.getFirstReadyUnit?.();
-        if (!id || typeof id.owner !== "number" || typeof id.id !== "number") return null;
-        const out = { owner: id.owner, id: id.id };
-        if (typeof id.type === "number") out.type = id.type;
-        return out;
-      }),
-    });
-  })()`;
-}
-
 function buildOperationValidationCommand(family: Civ7OperationFamily, input: Civ7OperationInput): string {
   return `(() => {
     ${operationRouterSource()}
@@ -5534,65 +5508,6 @@ function assertApproved(approval: Civ7ActionApproval, action: string): void {
   if (!approval || approval.approved !== true || !approval.reason.trim()) {
     throw new Civ7DirectControlError("command-failed", `Explicit approval with a reason is required before ${action}`);
   }
-}
-
-function isTurnCompletionAllowed(
-  status: Civ7TurnCompletionStatusResult,
-  fallbackPreflight?: Civ7PlayNotificationViewResult,
-): boolean {
-  if (probeValue(status.canEndTurn) === true) return true;
-  const cleanFallbackState = probeValue(status.hasSentTurnComplete) === false
-    && probeValue(status.blocker) === 0
-    && probeValue(status.firstReadyUnitId) === null;
-  if (!cleanFallbackState) return false;
-  if (fallbackPreflight === undefined) return false;
-  const blockingNotifications = fallbackPreflight.notifications.filter((notification) => notification.isEndTurnBlocking);
-  return blockingNotifications.every((notification) => isTurnCompletionFallbackNotification(notification, status));
-}
-
-function isTurnCompletionFallbackNotification(
-  notification: Civ7PlayNotificationSummary,
-  status: Civ7TurnCompletionStatusResult,
-): boolean {
-  const typeName = String(notification.typeName ?? "").toUpperCase();
-  if (notification.decision.category === "unit-command" && typeName.includes("COMMAND_UNITS")) {
-    return probeValue(status.blocker) === 0
-      && probeValue(status.firstReadyUnitId) === null
-      && notificationDetailsProveStaleCommandUnits(notification.details);
-  }
-  if (notification.decision.category === "informational-notification") {
-    return notification.canUserDismiss === true && isTurnCompletionFallbackInformationalType(typeName);
-  }
-  return false;
-}
-
-function notificationDetailsProveStaleCommandUnits(details: unknown): boolean {
-  if (!isRecord(details)) return false;
-  const enabledCloseoutCandidates = details.enabledCloseoutCandidates;
-  return details.kind === "unit-command-reconciliation"
-    && details.classification === "unit-command-stale-expired"
-    && details.staleExpiredWithoutEnabledCloseout === true
-    && Array.isArray(enabledCloseoutCandidates)
-    && enabledCloseoutCandidates.length === 0;
-}
-
-function isTurnCompletionFallbackInformationalType(typeName: string): boolean {
-  return typeName === "NOTIFICATION_UNIT_ATTACKED"
-    || typeName === "NOTIFICATION_DISTRICT_ATTACKED"
-    || typeName === "NOTIFICATION_RIVER_FLOODS_SEV0"
-    || typeName === "NOTIFICATION_RIVER_FLOODS_SEV1"
-    || typeName === "NOTIFICATION_RIVER_FLOODS_SEV2"
-    || typeName === "NOTIFICATION_STORM_ARRIVED"
-    || typeName === "NOTIFICATION_STORM_MOVED"
-    || typeName === "NOTIFICATION_STORM_DISSIPATED"
-    || typeName === "NOTIFICATION_VOLCANO_ACTIVE"
-    || typeName === "NOTIFICATION_VOLCANO_INACTIVE"
-    || typeName === "NOTIFICATION_VOLCANO_ERUPTS_SEV0"
-    || typeName === "NOTIFICATION_VOLCANO_ERUPTS_SEV1"
-    || typeName === "NOTIFICATION_VOLCANO_ERUPTS_SEV2"
-    || typeName === "NOTIFICATION_WONDER_COMPLETED"
-    || typeName === "NOTIFICATION_WONDER_FAILED"
-    || typeName === "NOTIFICATION_LEGACY_COMPLETED";
 }
 
 function autoplayConfigMatches(status: Civ7AutoplayStatusResult, options: Civ7AutoplayOptions): boolean {
